requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-pymongo
pydantic
flask-pydantic
flask-cors
psycopg2